requirements.txt
applescript==2021.2.9
pandas==1.3.4
requests==2.26.0
python-crontab==2.6.0
